Change port type

+++ models/bluetooth.c
@@ -63,7 +63,7 @@ void getCommand() {
 
 
 
-sdp_session_t *register_service(uint8_t rfcomm_channel) {
+sdp_session_t* register_service(int rfcomm_channel) {
 	uint32_t svc_uuid_int[] = { 0x38D46A42, 0xBDA14A6E, 0x9CD4CFC4, 0xB8056203 };
 	const char *service_name = "ArmBot Bluetooth server";
 	const char *svc_dsc = "An ArmBot server that interfaces with the ArmBot Android app";
